+++ packages/mhy-material-components/src/runtime/types/enums.d.ts
@@ -1,1 +1,1 @@
-type ThemeSetting = "auto" | "light" | "dark"
+// type ThemeSetting = "auto" | "light" | "dark"
