+++ packages/nuxt/components/character/CharacterCard.vue
