+++ packages/nuxt/components/character/CharacterIconCard.vue
@@ -1,6 +1,4 @@
 <script lang="ts" setup>
-import { useNuxtApp } from "#imports"
-
 defineProps<{
   to: string
   name: string
