+++ packages/nuxt/components/character/CharacterSelect.vue
@@ -1,6 +1,4 @@
 <script lang="ts" setup>
-import { computed, onMounted, ref, useRoute } from "#imports"
-
 interface Props {
   modelValue: string | undefined
   label?: string
